Reordering bayesian unit to lead with blackjack
This order creates implementation first and then brings the theory
later. I do this to both get the students "back into programming" after
the winter break and because the way I use bayesian learning in the
black jack game as a bit less than purely theoretical. Therefore, it's
easier to explain the hacky bayesian way I do it, then to build in the
theory later.